[backbone] Refactor common declarations into common.d.ts
